Web/LostFlag/includes/lostflag/requirements.txt
django
python-decouple
waitress
gunicorn[gevent]
whitenoise
requests
requests-async
redis